feat(ConnectivityService): extract BFS connectivity logic
refactor(GameService): delegate recomputeConnectivity to ConnectivityService
